requirements.txt
appdirs==1.4.4
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pipenv==2020.11.15
python-dotenv==0.15.0
pywin32==300
redis==3.5.3
requests==2.25.1
six==1.15.0
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
winclip32==0.6.0